+++ .github/CHANGELOG.md
@@ -3,9 +3,9 @@ This file contains update logs for this project. The top may contain a `Unreleas
 
 ## Unreleased
 ### Changed
+* Bungee/Velocity: Added `maintenance.singleserver.bypass.<server>` permission to give individual bypasses for proxied servers (`maintenance.bypass` will still bypass everything as the super-perm)
 * Added the messages of the help command into the language files
 * Added config option `kick-online-players` to disable the kicks when maintenance is enabled (new connections will still be blocked)... not sure why you'd want that, but you can :p
-* Bungee/Velocity: Added `maintenance.singleserver.bypass.<server>` permission to give individual bypasses for proxied servers (`maintenance.bypass` will still bypass everything as the super-perm)
 * The language file will now be automatically updated as well, so no need to delete and recreate it anymore, yay!
 * Updated missing and new translations in French language file (thanks to Spigot user Zendrique)
 * Updated missing and new translations in Spanish language file (thanks to Spigot user Vixo_Ulises)
@@ -13,7 +13,8 @@ This file contains update logs for this project. The top may contain a `Unreleas
 * Bungee/Velocity: Fixed the join-notifications being sent twice if both global and server maintenance are enabled, as well as when just a server has maintenance in some cases
 * Bungee: Fixed players being kicked *after* the initial connection and thus triggering a join and leave message before being kicked
 * Bungee: Fixed the kickmessage not being displayed in some cases
-* Paper: Fixed compatibility with ProtocolSupport: If running on Paper and having ProtocolSupport installed, Maintenance will switch to ProtocolLib's packet adapter, as PS does not fire Paper's ping event.
+* Paper: Fixed the playercount message not being displayed on Paper servers with ViaVersion
+* Paper: Fixed the motd not being changed at all on Paper servers with ProtocolSupport (-> ProtocolLib listener used, since ProtocolSupport does not fire Paper's ping event)
 * Fixed the config autoupdater when the config contained strings broken over multiple lines.
 ---
 ## 3.0.2 - May 17th 2019